+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [Unreleased]
 
+## [1.0.1] - 2026-08-06
+
 ### Fixed
 
 - Made instruction-bootstrap idempotency require exact source, destination, and